requirements.txt
numpy
arcade
pillow